Fix doctests
The order of Object keys can now be sorted for tests with a newtype
SortedDatum and new showArray and showObject functions. This avoids
having to adjust doctests in the future if the order of the Object
keys changes due to implementation details. All the necessary tests
have been adjusted accordingly.

Exclude http tests for now as httpbin.org appears to be unmaintained,
causing timeouts that make the tests fail.

Make sorted from the Doctext module more general by making it use
Functors instead of IO Datum specifically. This allows using it with
something like IO (Maybe Datum).

Adjust doctests for RethinkDB 2.4 that now includes write hooks.
Wildcard the new field for now to be able to run the tests against
both RethinkDB 2.4 as well as older versions.